+++ src/components/preloader/index.vue
@@ -1,8 +1,8 @@
 <template>
   <div>
-    <overlay :show="show" :transparent="true"></overlay>
-    <transition name="transition-modal">
-      <div class="preloader-modal" v-if="show">
+    <overlay :show="mutableShow" :transparent="true"></overlay>
+    <transition name="preloader-modal">
+      <div class="preloader-modal" v-if="mutableShow">
         <span class="preloader preloader-white"></span>
       </div>
     </transition>
@@ -19,9 +19,22 @@ export default {
   props: {
     show: {
       type: Boolean,
-      default: false,
-      twoWay: true,
-      required: true
+      default: false
+    }
+  },
+  data () {
+    return {
+      mutableShow: this.show
+    }
+  },
+  methods: {
+    open () {
+      this.mutableShow = true
+      this.$emit('open', this)
+    },
+    close () {
+      this.mutableShow = false
+      this.$emit('close', this)
     }
   }
 }